requirements.txt
# Broker - MetaTrader 5
MetaTrader5==5.0.4424

# IA - OpenAI
openai==1.30.1

# Notion - Log de operaciones
notion-client==2.2.1

# Utilidades
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
schedule==1.2.2
